+++ flowman-plugins/impala/src/main/scala/com/dimajix/flowman/spec/catalog/ImpalaCatalogSpec.scala
@@ -22,19 +22,19 @@ import com.dimajix.flowman.catalog.ExternalCatalog
 import com.dimajix.flowman.catalog.ImpalaExternalCatalog
 import com.dimajix.flowman.config.FlowmanConf
 import com.dimajix.flowman.execution.Context
-import com.dimajix.flowman.model.ConnectionIdentifier
 import com.dimajix.flowman.spec.annotation.CatalogType
+import com.dimajix.flowman.spec.connection.ConnectionReferenceSpec
 import com.dimajix.flowman.spec.connection.ImpalaConnection
 import com.dimajix.flowman.spec.connection.JdbcConnection
 
 
 @CatalogType(kind = "impala")
 class ImpalaCatalogSpec extends CatalogSpec {
-    @JsonProperty(value="connection", required=true) private var _connection:String = ""
+    @JsonProperty(value="connection", required=true) private var connection:ConnectionReferenceSpec = _
 
     override def instantiate(context: Context): ExternalCatalog = {
         val computeStats = context.flowmanConf.getConf(FlowmanConf.IMPALA_COMPUTE_STATS)
-        val con = context.getConnection(ConnectionIdentifier.parse(context.evaluate(_connection)))
+        val con = this.connection.instantiate(context).value
         val connection = con match {
             case jdbc:JdbcConnection => ImpalaExternalCatalog.Connection(
                 jdbc.url,