+++ flowman-spec/src/main/scala/com/dimajix/flowman/spec/relation/JdbcRelation.scala
@@ -16,12 +16,10 @@
 
 package com.dimajix.flowman.spec.relation
 
-import java.sql.Connection
 import java.sql.SQLInvalidAuthorizationSpecException
 import java.sql.SQLNonTransientConnectionException
 import java.sql.SQLNonTransientException
 import java.sql.Statement
-import java.util.Locale
 
 import scala.collection.mutable
 import scala.util.control.NonFatal
@@ -38,11 +36,6 @@ import org.slf4j.LoggerFactory
 import com.dimajix.common.SetIgnoreCase
 import com.dimajix.common.Trilean
 import com.dimajix.flowman.catalog.TableChange
-import com.dimajix.flowman.catalog.TableChange.AddColumn
-import com.dimajix.flowman.catalog.TableChange.DropColumn
-import com.dimajix.flowman.catalog.TableChange.UpdateColumnComment
-import com.dimajix.flowman.catalog.TableChange.UpdateColumnNullability
-import com.dimajix.flowman.catalog.TableChange.UpdateColumnType
 import com.dimajix.flowman.execution.Context
 import com.dimajix.flowman.execution.Execution
 import com.dimajix.flowman.execution.MigrationFailedException
@@ -55,16 +48,17 @@ import com.dimajix.flowman.jdbc.SqlDialect
 import com.dimajix.flowman.jdbc.SqlDialects
 import com.dimajix.flowman.jdbc.TableDefinition
 import com.dimajix.flowman.model.BaseRelation
-import com.dimajix.flowman.model.ConnectionIdentifier
+import com.dimajix.flowman.model.Connection
 import com.dimajix.flowman.model.PartitionField
 import com.dimajix.flowman.model.PartitionSchema
 import com.dimajix.flowman.model.PartitionedRelation
+import com.dimajix.flowman.model.Reference
 import com.dimajix.flowman.model.Relation
 import com.dimajix.flowman.model.ResourceIdentifier
 import com.dimajix.flowman.model.Schema
 import com.dimajix.flowman.model.SchemaRelation
+import com.dimajix.flowman.spec.connection.ConnectionReferenceSpec
 import com.dimajix.flowman.spec.connection.JdbcConnection
-import com.dimajix.flowman.types.Field
 import com.dimajix.flowman.types.FieldValue
 import com.dimajix.flowman.types.SingleValue
 import com.dimajix.flowman.types.{StructType => FlowmanStructType}
@@ -75,14 +69,13 @@ case class JdbcRelation(
     override val instanceProperties:Relation.Properties,
     override val schema:Option[Schema] = None,
     override val partitions: Seq[PartitionField] = Seq(),
-    connection: ConnectionIdentifier,
+    connection: Reference[Connection],
     properties: Map[String,String] = Map(),
     database: Option[String] = None,
     table: Option[String] = None,
     query: Option[String] = None
 ) extends BaseRelation with PartitionedRelation with SchemaRelation {
     private val logger = LoggerFactory.getLogger(classOf[JdbcRelation])
-    private lazy val jdbcConnection = context.getConnection(connection).asInstanceOf[JdbcConnection]
 
     def tableIdentifier : TableIdentifier = TableIdentifier(table.getOrElse(""), database)
 
@@ -323,7 +316,7 @@ case class JdbcRelation(
         }
     }
 
-    private def doCreate(con:Connection, options:JDBCOptions): Unit = {
+    private def doCreate(con:java.sql.Connection, options:JDBCOptions): Unit = {
         logger.info(s"Creating JDBC relation '$identifier', this will create JDBC table $tableIdentifier with schema\n${this.schema.map(_.treeString).orNull}")
         if (this.schema.isEmpty) {
             throw new UnspecifiedSchemaException(identifier)
@@ -412,7 +405,7 @@ case class JdbcRelation(
             }
         }
 
-        def alter(migrations:Seq[TableChange], con:Connection, options:JDBCOptions) : Unit = {
+        def alter(migrations:Seq[TableChange], con:java.sql.Connection, options:JDBCOptions) : Unit = {
             logger.info(s"Migrating JDBC relation '$identifier', this will alter JDBC table $tableIdentifier. New schema:\n${targetSchema.treeString}")
             if (migrations.isEmpty)
                 logger.warn("Empty list of migrations - nothing to do")
@@ -425,7 +418,7 @@ case class JdbcRelation(
             }
         }
 
-        def recreate(con:Connection, options:JDBCOptions) : Unit = {
+        def recreate(con:java.sql.Connection, options:JDBCOptions) : Unit = {
             try {
                 logger.info(s"Migrating JDBC relation '$identifier', this will recreate JDBC table $tableIdentifier. New schema:\n${targetSchema.treeString}")
                 JdbcUtils.dropTable(con, tableIdentifier, options)
@@ -463,7 +456,7 @@ case class JdbcRelation(
     }
 
     private def createProperties() : (String,Map[String,String]) = {
-        val connection = jdbcConnection
+        val connection = this.connection.value.asInstanceOf[JdbcConnection]
         val props = mutable.Map[String,String]()
         props.put(JDBCOptions.JDBC_URL, connection.url)
         props.put(JDBCOptions.JDBC_DRIVER_CLASS, connection.driver)
@@ -476,7 +469,7 @@ case class JdbcRelation(
         (connection.url,props.toMap)
     }
 
-    private def withConnection[T](fn:(Connection,JDBCOptions) => T) : T = {
+    private def withConnection[T](fn:(java.sql.Connection,JDBCOptions) => T) : T = {
         val (url,props) = createProperties()
         logger.debug(s"Connecting to jdbc source at $url")
 
@@ -532,7 +525,7 @@ case class JdbcRelation(
 
 
 class JdbcRelationSpec extends RelationSpec with PartitionedRelationSpec with SchemaRelationSpec {
-    @JsonProperty(value = "connection", required = true) private var connection: String = _
+    @JsonProperty(value = "connection", required = true) private var connection: ConnectionReferenceSpec = _
     @JsonProperty(value = "properties", required = false) private var properties: Map[String, String] = Map()
     @JsonProperty(value = "database", required = false) private var database: Option[String] = None
     @JsonProperty(value = "table", required = false) private var table: Option[String] = None
@@ -548,7 +541,7 @@ class JdbcRelationSpec extends RelationSpec with PartitionedRelationSpec with Sc
             instanceProperties(context),
             schema.map(_.instantiate(context)),
             partitions.map(_.instantiate(context)),
-            ConnectionIdentifier.parse(context.evaluate(connection)),
+            connection.instantiate(context),
             context.evaluate(properties),
             database.map(context.evaluate).filter(_.nonEmpty),
             table.map(context.evaluate).filter(_.nonEmpty),