+++ flowman-spec/src/main/scala/com/dimajix/flowman/spec/target/MergeTarget.scala
@@ -36,11 +36,11 @@ import com.dimajix.flowman.execution.Phase
 import com.dimajix.flowman.execution.VerificationFailedException
 import com.dimajix.flowman.graph.Linker
 import com.dimajix.flowman.model.BaseTarget
-import com.dimajix.flowman.model.IdentifierRelationReference
 import com.dimajix.flowman.model.MappingOutputIdentifier
 import com.dimajix.flowman.model.Reference
 import com.dimajix.flowman.model.Relation
 import com.dimajix.flowman.model.RelationIdentifier
+import com.dimajix.flowman.model.RelationReference
 import com.dimajix.flowman.model.ResourceIdentifier
 import com.dimajix.flowman.model.Target
 import com.dimajix.flowman.model.TargetInstance
@@ -54,7 +54,7 @@ object MergeTarget {
         new MergeTarget(
             Target.Properties(context),
             MappingOutputIdentifier(""),
-            IdentifierRelationReference(context, relation),
+            RelationReference(context, relation),
             conf.getConf(DEFAULT_TARGET_PARALLELISM),
             conf.getConf(DEFAULT_TARGET_REBALANCE)
         )